requirements.txt
diffusers==0.27.2
huggingface-hub==0.22.2
matplotlib==3.5.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
pytorch-lightning==2.2.1
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
torch==2.3.0
torchmetrics==1.3.2
torchvision==0.18.0
transformers==4.40.1
wandb==0.16.6
xformers==0.0.26.post1
lpips==0.1.4
accelerate==0.30.1
omegaconf==2.3.0
timm==1.0.3
lpips==0.1.4